initial work on refactoring adb
