+++ lib/devices/android/adb.js
@@ -129,29 +129,22 @@ ADB.prototype.checkSdkBinaryPresent = function(binary, cb) {
 ADB.prototype.checkAdbPresent = function(cb) {
   this.checkSdkBinaryPresent("adb", function(err, binaryLoc) {
     if (err) return cb(err);
+    if (this.adbCmd === this.adb) {
+      // if adbCmd is the same as default adb, update it too
+      this.adbCmd = binaryLoc.trim();
+    }
     this.adb = binaryLoc.trim();
     cb(null);
   }.bind(this));
 };
 
-ADB.prototype.checkAppPresent = function(cb) {
-  if (this.apkPath === null) {
-    logger.info("Not checking whether app is present since we are assuming " +
-                "it's already on the device");
-    cb(null);
-  } else {
-    logger.info("Checking whether app is actually present");
-    fs.stat(this.apkPath, function(err) {
-      if (err) {
-        logger.error("Could not find app apk at " + this.apkPath);
-        cb(new Error("Error locating the app apk, supposedly it's at " +
-                    this.apkPath + " but we can't stat it. Filesystem error " +
-                    "is " + err));
-      } else {
-        cb(null);
-      }
-    }.bind(this));
-  }
+ADB.prototype.exec = function(cmd, cb) {
+  cmd = this.adbCmd + ' ' + cmd;
+  logger.debug("executing: " + cmd);
+  exec(cmd, {maxBuffer: 524288}, function(err, stdout, stderr) {
+    if (err) logger.warn(err);
+    cb(err, stdout, stderr);
+  });
 };
 
 ADB.prototype.insertSelendroidManifest = function(serverPath, cb) {
@@ -435,238 +428,33 @@ ADB.prototype.checkApkCert = function(apk, cb) {
   });
 };
 
-ADB.prototype.getDeviceWithRetry = function(cb, count) {
+ADB.prototype.getDevicesWithRetry = function(timeoutMs, cb) {
+  if (typeof timeoutMs === "function") {
+    cb = timeoutMs;
+    timeoutMs = 20000;
+  }
+  var start = Date.now();
   logger.info("Trying to find a connected android device");
   var error = new Error("Could not find a connected Android device.");
-  var getDevices = function(innerCb) {
+  var getDevices = function() {
     this.getConnectedDevices(function(err, devices) {
-      if (typeof devices === "undefined" || devices.length === 0 || err) {
-        return innerCb(error);
-      }
-      innerCb(null);
-    });
-  }.bind(this);
-  getDevices(function(err) {
-    count = count || 1;
-    if (err) {
-      if (count < 10) {
-        logger.info("Could not find devices, restarting adb server...");
-        this.restartAdb(function() {
-          this.getDeviceWithRetry(cb, count + 1);
-        }.bind(this));
-      } else {
-        logger.info("Looked for devices " + count + " times. Giving up");
-        cb(error);
-      }
-    } else {
-      logger.info("Found device, no need to retry");
-      cb(null);
-    }
-  }.bind(this));
-};
-
-ADB.prototype.prepareDevice = function(onReady) {
-  logger.info("Preparing device for session");
-  async.series([
-    function(cb) { this.checkAppPresent(cb); }.bind(this),
-    function(cb) { this.checkAdbPresent(cb); }.bind(this),
-    function(cb) { this.prepareEmulator(cb); }.bind(this),
-    function(cb) { this.getDeviceWithRetry(cb);}.bind(this),
-    function(cb) { this.waitForDevice(cb); }.bind(this),
-    function(cb) { this.startLogcat(cb); }.bind(this)
-  ], onReady);
-};
-
-ADB.prototype.requestXmlCompression = function(cb) {
-  if (this.compressXml) {
-    this.android.proxy(["enableCompressedLayoutHeirarchy"], cb);
-  } else {
-    cb(null);
-  }
-};
-
-ADB.prototype.pushStrings = function(cb) {
-  var remotePath = '/data/local/tmp';
-  var stringsJson = 'strings.json';
-  if (!fs.existsSync(this.apkPath)) {
-   // apk doesn't exist locally so remove old strings.json
-   var pushCmd = this.adbCmd + ' shell rm ' + remotePath + '/' + stringsJson;
-   logger.debug("Apk doesn't exist. Removing old strings.json " + pushCmd);
-   exec(pushCmd, { maxBuffer: 524288 }, function(err, stdout, stderr) {
-     cb(null);
-   });
-  } else {
-    var stringsFromApkJarPath = path.resolve(helperJarPath,
-        'strings_from_apk.jar');
-    var outputPath = path.resolve(getTempPath(), this.appPackage);
-    var makeStrings = ['java -jar "', stringsFromApkJarPath,
-                       '" "', this.apkPath, '" "', outputPath, '"'].join('');
-    logger.debug(makeStrings);
-    exec(makeStrings, { maxBuffer: 524288 }, function(err, stdout, stderr) {
-      if (err) {
-        logger.debug(stderr);
-        return cb("error making strings");
-      }
-      var jsonFile = path.resolve(outputPath, stringsJson);
-
-      var pushCmd = this.adbCmd + ' push "' + jsonFile + '" ' + remotePath;
-      exec(pushCmd, { maxBuffer: 524288 }, function(err, stdout, stderr) {
-        cb(null);
-      });
-    }.bind(this));
-  }
-};
-
-ADB.prototype.startAppium = function(onReady, onExit) {
-  logger.info("Starting android appium");
-  this.onExit = onExit;
-
-  logger.debug("Using fast reset? " + this.fastReset);
-
-  async.series([
-    function(cb) { this.prepareDevice(cb); }.bind(this),
-    function(cb) { this.pushStrings(cb); }.bind(this),
-    function(cb) { this.requestXmlCompression(cb); }.bind(this),
-    function(cb) { this.uninstallApp(cb); }.bind(this),
-    function(cb) { this.installApp(cb); }.bind(this),
-    function(cb) { this.forwardPort(cb); }.bind(this),
-    function(cb) { this.pushAppium(cb); }.bind(this),
-    function(cb) { this.pushUnlock(cb); }.bind(this),
-    function(cb) { this.runBootstrap(cb, onExit); }.bind(this),
-    function(cb) { this.wakeUp(cb); }.bind(this),
-    function(cb) { this.unlockScreen(cb); }.bind(this),
-    function(cb) { this.startApp(cb); }.bind(this)
-  ], function(err, seriesInfo) {
-    onReady(err);
-  });
-};
-
-ADB.prototype.startChrome = function(onReady) {
-  logger.info("Starting Chrome");
-  logger.debug("Using fast reset? " + this.fastReset);
-
-  async.series([
-    function(cb) { this.prepareDevice(cb); }.bind(this),
-    function(cb) { this.installApp(cb); }.bind(this),
-    function(cb) { this.startApp(cb); }.bind(this)
-  ], onReady);
-};
-
-ADB.prototype.startSelendroid = function(serverPath, onReady) {
-  logger.info("Starting selendroid");
-  var modServerExists = false
-    , modAppPkg = this.appPackage + '.selendroid';
-  this.selendroidServerPath = path.resolve(getTempPath(),
-      'selendroid.' + this.appPackage + '.apk');
-
-  var checkModServerExists = function(cb) {
-    fs.stat(this.selendroidServerPath, function(err) {
-      modServerExists = !err;
-      cb();
-    });
-  }.bind(this);
-
-  var conditionalUninstallSelendroid = function(cb) {
-    if (!modServerExists) {
-      // assume we're going to rebuild selendroid and therefore
-      // need to uninstall it if it's already on device
-      logger.info("Rebuilt selendroid apk does not exist, uninstalling " +
-                  "any instances of it on device to make way for new one");
-      this.uninstallApk(modAppPkg, cb);
-    } else {
-      logger.info("Rebuilt selendroid apk exists, doing nothing");
-      cb();
-    }
-  }.bind(this);
-
-  var conditionalInsertManifest = function(cb) {
-    if (!modServerExists) {
-      logger.info("Rebuilt selendroid server does not exist, inserting " +
-                  "modified manifest");
-      this.insertSelendroidManifest(serverPath, cb);
-    } else {
-      logger.info("Rebuilt selendroid server already exists, no need to " +
-                  "rebuild it with a new manifest");
-      cb();
-    }
-  }.bind(this);
-
-  var conditionalInstallSelendroid = function(cb) {
-    this.checkAppInstallStatus(modAppPkg, function(e, installed) {
-      if (!installed) {
-        logger.info("Rebuilt selendroid is not installed, installing it");
-        this.installApk(this.selendroidServerPath, cb);
-      } else {
-        logger.info("Rebuilt selendroid is already installed");
-        cb();
-      }
-    }.bind(this));
-  }.bind(this);
-
-  async.series([
-    function(cb) { this.prepareDevice(cb); }.bind(this),
-    function(cb) { checkModServerExists(cb); },
-    function(cb) { conditionalUninstallSelendroid(cb); },
-    function(cb) { conditionalInsertManifest(cb); },
-    function(cb) { this.checkSelendroidCerts(this.selendroidServerPath, cb); }.bind(this),
-    function(cb) { conditionalInstallSelendroid(cb); },
-    function(cb) { this.installApp(cb); }.bind(this),
-    function(cb) { this.forwardPort(cb); }.bind(this),
-    function(cb) { this.pushUnlock(cb); }.bind(this),
-    function(cb) { this.unlockScreen(cb); }.bind(this),
-    function(cb) { this.pushSelendroid(cb); }.bind(this),
-    function(cb) { logger.info("Selendroid server is launching"); cb(); }
-  ], onReady);
-};
-
-ADB.prototype.pushSelendroid = function(cb) {
-  var activityString = this.appActivity;
-  if (activityString.indexOf(this.appPackage) === 0) {
-    activityString = activityString.substring(this.appPackage.length);
-  }
-  var cmd = this.adbCmd + " shell am instrument -e main_activity '" +
-            this.appPackage + activityString + "' " + this.appPackage +
-            ".selendroid/io.selendroid.ServerInstrumentation";
-  cmd = cmd.replace(/\.+/g,'.'); // Fix pkg..activity error
-  logger.info("Starting instrumentation process for selendroid with cmd: " +
-              cmd);
-  exec(cmd, { maxBuffer: 524288 }, function(err, stdout) {
-    if (err) return cb(err);
-    if (stdout.indexOf("Exception") !== -1) {
-      logger.error(stdout);
-      var msg = stdout.split("\n")[0] || "Unknown exception starting selendroid";
-      return cb(new Error(msg));
-    }
-    cb();
-  });
-};
-
-ADB.prototype.checkSelendroidCerts = function(serverPath, cb) {
-  var alreadyReturned = false
-    , checks = 0;
-
-  var onDoneSigning = function() {
-    checks++;
-    if (checks === 2 && !alreadyReturned) {
-      cb();
-    }
-  };
-
-  // these run in parallel
-  var apks = [serverPath, this.apkPath];
-  _.each(apks, function(apk) {
-    logger.info("Checking signed status of " + apk);
-    this.checkApkCert(apk, function(isSigned) {
-      if (isSigned) return onDoneSigning();
-      this.sign([apk], function(err) {
-        if (err && !alreadyReturned) {
-          alreadyReturned = true;
-          return cb(err);
+      if (err || devices.length < 1) {
+        if ((Date.now() - start) > timeoutMs) {
+          cb(error);
+        } else {
+          logger.info("Could not find devices, restarting adb server...");
+          setTimeout(function() {
+            this.restartAdb(function() {
+              getDevices();
+            }.bind(this));
+          }.bind(this), 1000);
         }
-        onDoneSigning();
-      });
-    }.bind(this));
-  }, this);
+      } else {
+        cb(null, devices);
+      }
+    });
+  }.bind(this);
+  getDevices();
 };
 
 ADB.prototype.getEmulatorPort = function(cb) {
@@ -689,6 +477,18 @@ ADB.prototype.getEmulatorPort = function(cb) {
   }.bind(this));
 };
 
+ADB.prototype.rimraf = function(path, cb) {
+   this.exec('shell rm -rf ' + path, cb);
+};
+
+ADB.prototype.push = function(localPath, remotePath, cb) {
+  try {
+    localPath = JSON.parse(localPath);
+  } catch(e) { }
+  localPath = JSON.stringify(localPath);
+  this.exec('push ' + localPath + ' ' + remotePath, cb);
+};
+
 ADB.prototype.getPortFromEmulatorString = function(emStr) {
   var portPattern = /emulator-(\d+)/;
   if (portPattern.test(emStr)) {
@@ -702,80 +502,53 @@ ADB.prototype.getRunningAVDName = function(cb) {
   this.sendTelnetCommand("avd name", cb);
 };
 
-ADB.prototype.prepareEmulator = function(cb) {
-  if (this.avdName !== null) {
-    this.getRunningAVDName(function(err, runningAVDName) {
-      if (!err && this.avdName.replace('@','') === runningAVDName) {
-        logger.info("Did not launch AVD because it was already running.");
-        cb(null);
-      } else {
-        logger.info("Launching Emulator with AVD " + this.avdName);
-        var killallCmd = isWindows ? "TASKKILL /IM emulator.exe" : "/usr/bin/killall -m emulator*";
-        exec(killallCmd, { maxBuffer: 524288 }, function(err, stdout) {
-          if (err) {
-            logger.info("Could not kill emulator. It was probably not running. : " + err.message);
-          }
-          this.checkSdkBinaryPresent("emulator", function(err, emulatorBinaryPath) {
-            if (err) {
-              return cb(err);
-            }
-            if (this.avdName[0] !== "@") {
-              this.avdName = "@" + this.avdName;
-            }
-            var emulatorProc = spawn(emulatorBinaryPath.substr(1, emulatorBinaryPath.length - 2), [this.avdName]);
-            var timeoutMs = 120000;
-            var now = Date.now();
-            var checkEmulatorAlive = function() {
-              this.restartAdb(function() {
-                this.getConnectedDevices(function(err, devices) {
-                  if (!err && devices.length) {
-                    cb(null, true);
-                  } else if (Date.now() < (now + timeoutMs)) {
-                    setTimeout(checkEmulatorAlive, 2000);
-                  } else {
-                    cb(new Error("Emulator didn't come up in " + timeoutMs + "ms"));
-                  }
-                }.bind(this));
-              }.bind(this));
-            }.bind(this);
-            checkEmulatorAlive();
-          }.bind(this));
-        }.bind(this));
-      }
-    }.bind(this));
-  } else {
+ADB.prototype.killAllEmulators = function(cb) {
+  var killallCmd = isWindows ?
+    "TASKKILL /IM emulator.exe" :
+    "/usr/bin/killall -m emulator*";
+  exec(killallCmd, { maxBuffer: 524288 }, function(err) {
+    if (err) {
+      logger.info("Could not kill emulator. It was probably not running. : " + err.message);
+    }
     cb();
-  }
+  });
 };
 
+ADB.prototype.launchAVD = function(avdName, cb) {
+  logger.info("Launching Emulator with AVD " + avdName);
+  this.checkSdkBinaryPresent("emulator", function(err, emulatorBinaryPath) {
+    if (err) return cb(err);
+
+    if (avdName[0] !== "@") {
+      avdName = "@" + avdName;
+    }
+
+    spawn(emulatorBinaryPath.substr(1, emulatorBinaryPath.length - 2),
+      [avdName]);
+    this.getDeviceWithRetry(120000, cb);
+  }.bind(this));
+};
 
 ADB.prototype.getConnectedDevices = function(cb) {
   this.debug("Getting connected devices...");
-  exec(this.adb + " devices", { maxBuffer: 524288 }, function(err, stdout) {
-    if (err) {
-      logger.error(err);
-      cb(err);
-    } else if (stdout.toLowerCase().indexOf("error") !== -1) {
+  this.exec("devices", function(err, stdout) {
+    if (err) return cb(err);
+    if (stdout.toLowerCase().indexOf("error") !== -1) {
       logger.error(stdout);
       cb(new Error(stdout));
     } else {
       var devices = [];
       _.each(stdout.split("\n"), function(line) {
-        if (line.trim() !== "" && line.indexOf("List of devices") === -1 && line.indexOf("* daemon") === -1 && line.indexOf("offline") == -1) {
+        if (line.trim() !== "" &&
+            line.indexOf("List of devices") === -1 &&
+            line.indexOf("* daemon") === -1 &&
+            line.indexOf("offline") == -1) {
           var lineInfo = line.split("\t");
-          devices.push({udid: lineInfo[0], state: lineInfo[1]}); // state is either "device" or "offline", afaict
+          // state is either "device" or "offline", afaict
+          devices.push({udid: lineInfo[0], state: lineInfo[1]});
         }
       });
       this.debug(devices.length + " device(s) connected");
-      if (devices.length) {
-        this.debug("Setting device id to " + (this.udid || devices[0].udid));
-        this.emulatorPort = null;
-        var emPort = this.getPortFromEmulatorString(devices[0].udid);
-        this.setDeviceId(this.udid || devices[0].udid);
-        if (emPort && !this.udid) {
-          this.emulatorPort = emPort;
-        }
-      }
       cb(null, devices);
     }
   }.bind(this));
@@ -1006,6 +779,10 @@ ADB.prototype.setDeviceId = function(deviceId) {
   this.adbCmd = this.adb + " -s " + deviceId;
 };
 
+ADB.prototype.setEmulatorPort = function(emPort) {
+  this.emulatorPort = emPort;
+};
+
 ADB.prototype.requireDeviceId = function() {
   if (!this.curDeviceId) {
     throw new Error("This method requires that a device ID is set. " +
@@ -1035,11 +812,9 @@ ADB.prototype.requireApk = function() {
 
 ADB.prototype.waitForDevice = function(cb) {
   var doWait = function(innerCb) {
-    this.debug("Waiting for device " + this.curDeviceId + " to be ready " +
-               "and to respond to shell commands (timeout = " +
-               this.appDeviceReadyTimeout + ")");
+    this.debug("Waiting for device to be ready and to respond to shell " +
+               "commands (timeout = " + this.appDeviceReadyTimeout + ")");
     var movedOn = false
-      , cmd = this.adbCmd + " wait-for-device"
       , timeoutSecs = parseInt(this.appDeviceReadyTimeout, 10);
 
     setTimeout(function() {
@@ -1050,21 +825,21 @@ ADB.prototype.waitForDevice = function(cb) {
       }
     }.bind(this), timeoutSecs * 1000);
 
-    exec(cmd, { maxBuffer: 524288 }, function(err) {
+    this.exec("wait-for-device", function(err) {
       if (!movedOn) {
         if (err) {
           logger.error("Error running wait-for-device");
           movedOn = true;
           innerCb(err);
         } else {
-          exec(this.adbCmd + " shell echo 'ready'", function(err) {
+          this.exec("shell echo 'ready'", function(err) {
             if (!movedOn) {
               movedOn = true;
               if (err) {
                 logger.error("Error running shell echo: " + err);
                 innerCb(err);
               } else {
-                innerCb(null);
+                innerCb();
               }
             }
           }.bind(this));
@@ -1317,8 +1092,7 @@ ADB.prototype.waitForActivity = function(cb, waitMsOverride) {
 
 ADB.prototype.uninstallApk = function(pkg, cb) {
   logger.info("Uninstalling " + pkg);
-  var cmd = this.adbCmd + " uninstall " + pkg;
-  exec(cmd, { maxBuffer: 524288 }, function(err, stdout) {
+  this.exec("uninstall " + pkg, function(err, stdout) {
     if (err) {
       logger.error(err);
       cb(err);
@@ -1329,7 +1103,7 @@ ADB.prototype.uninstallApk = function(pkg, cb) {
       } else {
         logger.debug("App was not uninstalled, maybe it wasn't on device?");
       }
-      cb(null);
+      cb();
     }
   });
 };
@@ -1397,50 +1171,28 @@ ADB.prototype.removeOldApks = function(cb) {
   ], function(err, matchingApkFound) { cb(null, matchingApkFound); });
 };
 
-// This is only invoked after checking if the app is installed.
-// installAppApk will always install the apk.
-ADB.prototype.installAppApk = function(apk, cb) {
+ADB.prototype.installApp = function(apk, cb) {
+  var remotePath = '';
+
   var getMD5 = function(cb) {
     fs.readFile(apk, function(err, buffer) {
       this.appMD5 = md5(buffer);
-      cb(null);
+      remotePath = '/data/local/tmp/' + this.appMD5 + '.apk';
+      cb();
     }.bind(this));
   }.bind(this);
 
-  var adbMakeFolder = function(cb) {
-    var cmd = this.adbCmd + ' shell "mkdir /data/local/tmp/"';
-    logger.info("adbMakeFolder: " + cmd);
-    exec(cmd, { maxBuffer: 524288 }, function(err, stdout) {
-      cb(null);
-    });
+  var makeFolder = function(cb) {
+    this.exec('shell "mkdir /data/local/tmp/"', cb);
   }.bind(this);
 
-  var adbPush = function(cb) {
-    var cmd = this.adbCmd + ' push "' + apk + '" "/data/local/tmp/' + this.appMD5 + '.apk"';
-    logger.info("adbPush: " + cmd);
-    exec(cmd, { maxBuffer: 524288 }, function(err, stdout) {
-      if (err) {
-        logger.error(err);
-        cb(err);
-      } else {
-        logger.debug(stdout);
-        cb(null);
-      }
-    });
+  var pushApp = function(cb) {
+    this.push(apk, remotePath, cb);
   }.bind(this);
 
-  var adbInstall = function(cb) {
-    var cmd = this.adbCmd + ' shell "pm install -r /data/local/tmp/' + this.appMD5 + '.apk"';
-    logger.info("adbInstall: " + cmd);
-    exec(cmd, { maxBuffer: 524288 }, function(err, stdout) {
-      if (err) {
-        logger.error(err);
-        cb(err);
-      } else {
-        logger.debug(stdout);
-        cb(null);
-      }
-    });
+  var installFromDevice = function(cb) {
+    var cmd = 'shell "pm install -r ' + remotePath + '"';
+    this.exec(cmd, cb);
   }.bind(this);
 
   var signLocalApk = function(cb) {
@@ -1448,45 +1200,24 @@ ADB.prototype.installAppApk = function(apk, cb) {
       if (!appSigned) {
         this.sign([apk], cb);
       } else {
-        cb(null);
+        cb();
       }
     }.bind(this));
   }.bind(this);
 
   async.waterfall([
     function(cb) { signLocalApk(cb); },
-    function(cb) { adbMakeFolder(cb); },
+    function(cb) { makeFolder(cb); },
     function(cb) { getMD5(cb); },
     function(cb) { this.removeOldApks(cb); }.bind(this),
     // If the apk is already on device, then don't push it again.
     function(matchingApkFound, cb) {
-      if (matchingApkFound) { cb(null); } else { adbPush(cb); }
+      if (matchingApkFound) { cb(null); } else { pushApp(cb); }
     },
     function(cb) {
-      adbInstall(cb);
+      installFromDevice(cb);
     }
   ], cb);
-};
-
-ADB.prototype.uninstallApp = function(cb) {
-  var next = function() {
-    this.requireDeviceId();
-    this.requireApp();
-    this.debug("Uninstalling app " + this.appPackage);
-
-    this.uninstallApk(this.appPackage, function(err) {
-      if (err) return cb(err);
-      cb(null);
-    }.bind(this));
-  }.bind(this);
-
-  if (this.skipUninstall) {
-    this.debug("Not uninstalling app since server not started with " +
-               "--full-reset");
-    cb();
-  } else {
-    next();
-  }
 };
 
 ADB.prototype.runFastReset = function(cb) {
@@ -1515,59 +1246,18 @@ ADB.prototype.runFastReset = function(cb) {
   ], cb);
 };
 
-ADB.prototype.checkAppInstallStatus = function(pkg, cb) {
+ADB.prototype.isAppInstalled = function(pkg, cb) {
   var installed = false;
-  this.requireDeviceId();
 
   logger.debug("Getting install status for " + pkg);
-  var listPkgCmd = this.adbCmd + " shell pm list packages -3 " + pkg;
-  exec(listPkgCmd, { maxBuffer: 524288 }, function(err, stdout) {
+  var listPkgCmd = "shell pm list packages -3 " + pkg;
+  this.exec(listPkgCmd, function(err, stdout) {
+    if (err) return cb(err);
     var apkInstalledRgx = new RegExp('^package:' +
         pkg.replace(/([^a-zA-Z])/g, "\\$1") + '$', 'm');
     installed = apkInstalledRgx.test(stdout);
     cb(null, installed);
   });
-};
-
-ADB.prototype.installApp = function(cb) {
-  var installApp = false;
-  this.requireDeviceId();
-
-  if (this.apkPath === null) {
-    logger.info("Not installing app since we launched with a package instead " +
-                "of an app path");
-    return cb(null);
-  }
-
-  this.requireApk();
-
-  var determineInstallStatus = function(cb) {
-    logger.info("Determining app install");
-    this.checkAppInstallStatus(this.appPackage, function(err, installed) {
-      installApp = !installed;
-      cb();
-    });
-  }.bind(this);
-
-  var doInstall = function(cb) {
-    if (installApp) {
-      this.debug("Installing app apk");
-      this.installAppApk(this.apkPath, cb);
-    } else { cb(null); }
-  }.bind(this);
-
-  var doFastReset = function(cb) {
-    // App is already installed so reset it.
-    if (!installApp && this.fastReset) {
-      this.runFastReset(cb);
-    } else { cb(null); }
-  }.bind(this);
-
-  async.series([
-    function(cb) { determineInstallStatus(cb); },
-    function(cb) { doInstall(cb); },
-    function(cb) { doFastReset(cb); }
-  ], cb);
 };
 
 ADB.prototype.back = function(cb) {
