+++ lib/devices/android/android.js
@@ -1,11 +1,17 @@
 "use strict";
 
 var errors = require('../../server/errors.js')
+  , exec = require('child_process').exec
+  , path = require('path')
+  , fs = require('fs')
   , adb = require('./adb.js')
+  , helpers = require('../../helpers.js')
+  , getTempPath = helpers.getTempPath
   , _ = require('underscore')
   , logger = require('../../server/logger.js').get('appium')
   , deviceCommon = require('../common.js')
   , status = require("../../server/status.js")
+  , helperJarPath = path.resolve(__dirname, 'helpers')
   , async = require('async')
   , androidController = require('./android-controller.js')
   , UnknownError = errors.UnknownError;
@@ -143,14 +149,91 @@ Android.prototype.start = function(cb, onDie) {
   }
 };
 
+// XXX convert to use new adb
 Android.prototype.startAppium = function(onLaunch, onExit) {
-  this.adb.startAppium(onLaunch, onExit);
+  logger.info("Starting android appium");
+  this.adb.onExit = onExit;
+
+  logger.debug("Using fast reset? " + this.fastReset);
+
+  async.series([
+    function(cb) { this.prepareDevice(cb); }.bind(this),
+    function(cb) { this.pushStrings(cb); }.bind(this),
+    function(cb) { this.requestXmlCompression(cb); }.bind(this),
+    function(cb) { this.uninstallApp(cb); }.bind(this),
+    function(cb) { this.installApp(cb); }.bind(this),
+    function(cb) { this.forwardPort(cb); }.bind(this),
+    function(cb) { this.pushAppium(cb); }.bind(this),
+    function(cb) { this.pushUnlock(cb); }.bind(this),
+    function(cb) { this.runBootstrap(cb, onExit); }.bind(this),
+    function(cb) { this.wakeUp(cb); }.bind(this),
+    function(cb) { this.unlockScreen(cb); }.bind(this),
+    function(cb) { this.startApp(cb); }.bind(this)
+  ], function(err) {
+    onLaunch(err);
+  });
+};
+
+Android.prototype.pushStrings = function(cb) {
+  var remotePath = '/data/local/tmp';
+  var stringsJson = 'strings.json';
+  if (!fs.existsSync(this.apkPath)) {
+    // apk doesn't exist locally so remove old strings.json
+    logger.debug("Apk doesn't exist. Removing old strings.json");
+    this.adb.rimraf(remotePath + '/' + stringsJson, function(err) {
+      if (err) return cb(new Error("Could not remove old strings"));
+      cb();
+    });
+  } else {
+    var stringsFromApkJarPath = path.resolve(helperJarPath,
+        'strings_from_apk.jar');
+    var outputPath = path.resolve(getTempPath(), this.appPackage);
+    var makeStrings = ['java -jar "', stringsFromApkJarPath,
+                       '" "', this.apkPath, '" "', outputPath, '"'].join('');
+    logger.debug(makeStrings);
+    exec(makeStrings, { maxBuffer: 524288 }, function(err, stdout, stderr) {
+      if (err) {
+        logger.debug(stderr);
+        return cb(new Error("Could not make strings from apk"));
+      }
+      var jsonFile = path.resolve(outputPath, stringsJson);
+      this.adb.push(jsonFile, remotePath, function(err) {
+        if (err) return cb(new Error("Could not push strings.json"));
+        cb();
+      });
+    }.bind(this));
+  }
 };
 
 Android.prototype.timeoutWaitingForCommand = function() {
   logger.info("Didn't get a new command in " + (this.commandTimeoutMs / 1000) +
               " secs, shutting down...");
   this.stop();
+};
+
+Android.prototype.requestXmlCompression = function(cb) {
+  if (this.compressXml) {
+    this.proxy(["enableCompressedLayoutHeirarchy"], cb);
+  } else {
+    cb();
+  }
+};
+
+Android.prototype.uninstallApp = function(cb) {
+  var next = function() {
+    this.adb.uninstallApk(this.appPackage, function(err) {
+      if (err) return cb(err);
+      cb(null);
+    }.bind(this));
+  }.bind(this);
+
+  if (this.skipUninstall) {
+    this.debug("Not uninstalling app since server not started with " +
+               "--full-reset");
+    cb();
+  } else {
+    next();
+  }
 };
 
 Android.prototype.stop = function(cb) {
