+++ lib/devices/android/selendroid.js
@@ -1,10 +1,8 @@
 "use strict";
 
-var errors = require('../../server/errors.js')
-  , adb = require('./adb.js')
+var adb = require('./adb.js')
   , _ = require('underscore')
   , deviceCommon = require('../common.js')
-  , request = deviceCommon.request
   , proxyTo = deviceCommon.proxyTo
   , getLog = deviceCommon.getLog
   , getLogTypes = deviceCommon.getLogTypes
@@ -12,8 +10,7 @@ var errors = require('../../server/errors.js')
   , status = require("../../server/status.js")
   , fs = require('fs')
   , async = require('async')
-  , path = require('path')
-  , UnknownError = errors.UnknownError;
+  , path = require('path');
 
 var Selendroid = function(opts) {
   this.opts = opts;
@@ -46,6 +43,127 @@ Selendroid.prototype.start = function(cb) {
     function(cb) { this.createSession(cb); }.bind(this),
   ], cb);
 };
+
+// XXX convert to new way of using adb
+ADB.prototype.startSelendroid = function(serverPath, onReady) {
+  logger.info("Starting selendroid");
+  var modServerExists = false
+    , modAppPkg = this.appPackage + '.selendroid';
+  this.selendroidServerPath = path.resolve(getTempPath(),
+      'selendroid.' + this.appPackage + '.apk');
+
+  var checkModServerExists = function(cb) {
+    fs.stat(this.selendroidServerPath, function(err) {
+      modServerExists = !err;
+      cb();
+    });
+  }.bind(this);
+
+  var conditionalUninstallSelendroid = function(cb) {
+    if (!modServerExists) {
+      // assume we're going to rebuild selendroid and therefore
+      // need to uninstall it if it's already on device
+      logger.info("Rebuilt selendroid apk does not exist, uninstalling " +
+                  "any instances of it on device to make way for new one");
+      this.uninstallApk(modAppPkg, cb);
+    } else {
+      logger.info("Rebuilt selendroid apk exists, doing nothing");
+      cb();
+    }
+  }.bind(this);
+
+  var conditionalInsertManifest = function(cb) {
+    if (!modServerExists) {
+      logger.info("Rebuilt selendroid server does not exist, inserting " +
+                  "modified manifest");
+      this.insertSelendroidManifest(serverPath, cb);
+    } else {
+      logger.info("Rebuilt selendroid server already exists, no need to " +
+                  "rebuild it with a new manifest");
+      cb();
+    }
+  }.bind(this);
+
+  var conditionalInstallSelendroid = function(cb) {
+    this.checkAppInstallStatus(modAppPkg, function(e, installed) {
+      if (!installed) {
+        logger.info("Rebuilt selendroid is not installed, installing it");
+        this.installApk(this.selendroidServerPath, cb);
+      } else {
+        logger.info("Rebuilt selendroid is already installed");
+        cb();
+      }
+    }.bind(this));
+  }.bind(this);
+
+  async.series([
+    function(cb) { this.prepareDevice(cb); }.bind(this),
+    function(cb) { checkModServerExists(cb); },
+    function(cb) { conditionalUninstallSelendroid(cb); },
+    function(cb) { conditionalInsertManifest(cb); },
+    function(cb) { this.checkSelendroidCerts(this.selendroidServerPath, cb); }.bind(this),
+    function(cb) { conditionalInstallSelendroid(cb); },
+    function(cb) { this.installApp(cb); }.bind(this),
+    function(cb) { this.forwardPort(cb); }.bind(this),
+    function(cb) { this.pushUnlock(cb); }.bind(this),
+    function(cb) { this.unlockScreen(cb); }.bind(this),
+    function(cb) { this.pushSelendroid(cb); }.bind(this),
+    function(cb) { logger.info("Selendroid server is launching"); cb(); }
+  ], onReady);
+};
+
+// XXX convert to new way of using adb
+ADB.prototype.pushSelendroid = function(cb) {
+  var activityString = this.appActivity;
+  if (activityString.indexOf(this.appPackage) === 0) {
+    activityString = activityString.substring(this.appPackage.length);
+  }
+  var cmd = this.adbCmd + " shell am instrument -e main_activity '" +
+            this.appPackage + activityString + "' " + this.appPackage +
+            ".selendroid/io.selendroid.ServerInstrumentation";
+  cmd = cmd.replace(/\.+/g,'.'); // Fix pkg..activity error
+  logger.info("Starting instrumentation process for selendroid with cmd: " +
+              cmd);
+  exec(cmd, { maxBuffer: 524288 }, function(err, stdout) {
+    if (err) return cb(err);
+    if (stdout.indexOf("Exception") !== -1) {
+      logger.error(stdout);
+      var msg = stdout.split("\n")[0] || "Unknown exception starting selendroid";
+      return cb(new Error(msg));
+    }
+    cb();
+  });
+};
+
+// XXX convert to new way of using adb
+ADB.prototype.checkSelendroidCerts = function(serverPath, cb) {
+  var alreadyReturned = false
+    , checks = 0;
+
+  var onDoneSigning = function() {
+    checks++;
+    if (checks === 2 && !alreadyReturned) {
+      cb();
+    }
+  };
+
+  // these run in parallel
+  var apks = [serverPath, this.apkPath];
+  _.each(apks, function(apk) {
+    logger.info("Checking signed status of " + apk);
+    this.checkApkCert(apk, function(isSigned) {
+      if (isSigned) return onDoneSigning();
+      this.sign([apk], function(err) {
+        if (err && !alreadyReturned) {
+          alreadyReturned = true;
+          return cb(err);
+        }
+        onDoneSigning();
+      });
+    }.bind(this));
+  }, this);
+};
+
 
 Selendroid.prototype.stop = function(cb) {
   logger.info("Stopping selendroid server");
